rename http.ts to nghttp.ts to avoid resolution colission with node's http module
